Fix titlebar visibility with virtual monitors
Virtual monitors and monitors are now a single page, so checking
the popcount is not enough - it'll return 2 if VIRTUAL | MONITOR
was passed, but in fact only a single page is visible.

Fix that by checking if a WINDOW type is passed to, which will
in fact cause two pages of the stack to be visible.