screenshotdialog: Skip dialog for non-interactive screenshots
But only when the screenshot portal is version 3 or greater.
Before version 3, no portal-sided permissions were asked for
non-interactive screenshots, so we still have to show some
indication that the screen content was grabbed; but starting
from version 3, xdg-desktop-portal queries the user using the
access portal, so we don't have to do it again.