screenshot: Implement version 2
xdg-desktop-portal 1.15.0 has an update to the Screenshot portal
D-Bus API for frontends, with a version bump. This new version
also contains a behavioural change where, if the frontend implements
version 2, it'll request the 'screenshot' permission. This allows
us to skip the permission dialog in this case.

Implement version 2 of org.freedesktop.portal.impl.Screenshot, and
skip the screenshot dialog when we know that a permission has been
granted, and when the screenshot is non-interactive.