filechooser: Remember last app folder
The GNOME implementation of the FileChooser portal doesn't make any
effort to remember the last folder of any given app. This can be
quite frustrating, and a significant usability problem, especially
when doing repetitive operations on the same app that require picking
files over and over.

Remembering the last folder selected by a given app is a quick win
that doesn't require too much introspection design-wise, so let's
just do it.