screencast: Simplify feature advertisement
Back in the day, when xdg-desktop-portal-gnome was actually
xdg-desktop-portal-gtk, Mutter and x-d-p-gtk releases were out of
sync, which forced it to runtime check for the Mutter Screencast
D-Bus API version in order to advertise the correct set of features
to the portal frontend.

This changed when we moved it to be GNOME-specific, and bound to
GNOME release schedules. Now it is unreasonable to expect that e.g.
xdg-desktop-portal-gnome 43 is running on top of Mutter 42. That
means we can make assumptions about versioning in the code.

Cleanup the feature advertisement code in screencast.c to always,
unconditionally advertise the current features, since Mutter 43
supports them all.