externalwindow: Handle empty window handle gracefully again
Some applications will not provide a parent window, meaning we'll have
nothing to create our external window object from. We still attempted
that, resulting in an attempt to use the NULL as the xdg_foreign parent,
which is not valid and results in a Wayland protocol error.

While at it, also guard against bogus handle prefixes in the Wayland
external window implementation.

This will make e.g. the screen sharing dialog opened from Chromium work
again.