screencastwidget: Always allow transient persist modes
Transient persist modes are mainly used by browsers that need to
restore a session across different windows, but don't really need
eternal permission to access a particular monitor.

When screencasting, GNOME Shell already shows an indicator, so we
don't really need to keep asking users for the permission to
restore it.

Hide the "Remember this selection" check when the app asks for
persistent permission.