screencast: Add monitor selection container
Previously, it was difficult to determine which monitor is which,
especially, with identical display names.

Now, a `ScreenCastGeometryContainer` is used to layout `GtkToggleButtons`
in a geometry which corresponds to the logical monitor layout.
The old monitor selection `GtkListBox` has been removed.

Additionally, the stack switcher for the source type
has been moved inside the dialog window.

Closes: https://gitlab.gnome.org/GNOME/xdg-desktop-portal-gnome/-/issues/9