Use ci-templates for image building
This will avoid the base image used for building and in the future
testing potentially changing every pipeline, it will avoid needing to
download and install build dependencies every run, while still making it
convenient to change the contents of the images.

The meson install script is copied from mutter.