remotedesktop: Add support for persistent sessions
This works roughly the same as persistent screen cast sessions, i.e. by
serializing the state (what streams and what device types) into a
variant that is then sent to the portal, which turns it into a token
where the variant used for restoring is stored in the permission store.

The logic surrounding saving and restoring streams is shared with the
screen cast portal, and the remote desktop side only deals with a
bitmask of shared device types, as well as creation and last use time
stamps.