Expose settings backend when running outside of GNOME
This means it'll no longer exit when launched in an incompatible
environment, but will continue but only exposing the
org.freedesktop.impl.portal.Settings interface, and none of the other
portals.

This is needed to fulfill the expectations of xdg-desktop-portal, who
will unconditionally launch every portal backend and merge all the
settings from each and every portal backend into a single look up table.

Fetching the session bus is moved earlier, as otherwise it might try to
reuse the failed attempt from trying to initialize GTK.