screencastwidget: Use shellintrospect internal list model
Rather than trying to recreate what shellintrospect is doing,
simply attach the listmodel to the listbox. Use a filter list
model to filter out some of the windows as we were previously
manually doing.

Besides the code benefits, this fixes a user visible issue
where the list completely recreates on any window changes.
This is particularly problematic when a window is selected
(and becomes unselected) on the window change.