gitlab-ci.yml: Fix the Ubuntu CI adwaita version
In Ubuntu build test, xdg-desktop-portal version being installed
is 1.16.0, when at least version 1.17.0 is required to build
this repo. With current 1.16.0 version, build fails, and so CI
build test also fails.

This MR changes the xdg-desktop-portal version being downloaded
to 1.18.0, which is the one being shipped in Ubuntu Mantic (the
ubuntu version used in the build test). This change fixes this
bug.