access: Add new portal implementation
Right now the portal is implemented by gnome-shell as a system-modal
dialog.

However when we have a parent window to attach to, using a regular
modal dialog is less disruptive, and provides a clearer link between
permission request and requesting app.

So provide a simple portal implementation that uses a message dialog
if we have a parent window, and delegates the request to gnome-shell
otherwise.