screencastwidget: Stop relying on GtkListBox selecting widgets
It gets too much in the way, and makes it hard to impossible to
implement click-to-toggle as well as not resetting the selection when
using keyboard navigation.

Instead rely on the GtkListBox cursor to navigate, and track selection
state ourself. A tiny CSS class is added to deal with making it look the
same as before.

Closes: https://gitlab.gnome.org/GNOME/xdg-desktop-portal-gnome/-/issues/93