filechooser: Delegate implementation to nautilus
So far we have been using GtkFileChooserDialog to provide the UI for
the FileChooser portal, as inherited from xdg-desktop-portal-gtk.

But GtkFileChooserDialog is deprecated and for various other reasons
nautilus is now providing a FileChooser portal implementation.

Ideally, nautilus could provide the implementation directly and we
could drop ours. Unfortunately, due to a bug around the color-scheme
setting involving portals in libadwaita, if nautilus itself presents
is declared as a portal implementation with a .portal file, then it
will not follow the dark mode setting.

So, for the time being, we still act as the backend implementing the
FileChooser portal, but relay messages into/from nautilus instead,
almost unchanged.

"Almost" because, for convenience, we keep the responsibility of
saving and restoring the previously used folders for each app. We
also keep the responsibility of adding files to recents for now.

Part of: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3401