display-state-tracker: Use proxy from source object in callback
The proxy in our instance struct might have been replaced or cleared,
but the source object still have a ref on the old proxy, and is the
correct proxy to finish the call on. Using the source object fixes a
crash when the proxy clears before the method call finishes.

Closes: https://gitlab.gnome.org/GNOME/xdg-desktop-portal-gnome/-/issues/141