Initialize GTK before opening display for X11, too
Commit 73e94e68 "externalwindow: Call gtk_init before opening a display"
only acted on the Wayland code path. For those still using the Xorg
session, we need a similar change in the X11 code path to avoid crashes
after upgrading to GTK 4.17.x.

This is unnecessary on main since it's redundant with commit c04aea5f
"Outsource external window management to libgxdp", but that commit is
unsuitable for v47.x since it adds a hard dependency on GTK 4.17.x.

[smcv: Add commit message]
Bug-Debian: https://bugs.debian.org/1096094