notification: Fix build against xdg-desktop-portal >= 1.19.1
Based on patches by Georges Basile Stavracas Neto (in x-d-p-gtk) and
Sébastien Noel (on https://bugs.debian.org/1096094).

This is unnecessary on main because it's redundant with commit 4cb1575c
"deps: Upgrade portal to >=1.19.1", but that commit is unsuitable for
backport to 47.x since it adds a hard dependency on the newer version.

Bug-Debian: https://bugs.debian.org/1096094
Signed-off-by: Simon McVittie <smcv@debian.org>