background: Avoid parsing the non-sandboxed-apps desktop files
xdg-desktop-portal-gnome is targetting only GNOME and thus the shell and
here we're always exposing the sandboxed-app-id property for all the
running applications.

As per this, there's no need to use the fallback path for checking all
the other app desktop files to check if they are Flatpak applications.

Since they won't be, or if they are, they might be hand-crafted desktop
files. While we should rely on the window PID as mutter does.

Part-of: <https://gitlab.gnome.org/GNOME/xdg-desktop-portal-gnome/-/merge_requests/226>