printing: Do not leak print dialog
Currently, print dialog polls printers even when it was already closed.
This happens because it is not finalized and backends are not disposed.
This commit fixes the situation by unref-ing print dialog at the end
of the function where it was created since required references were already
created.

Part-of: <https://gitlab.gnome.org/GNOME/xdg-desktop-portal-gnome/-/merge_requests/241>